requirements.txt
aiohttp==3.3.2
async-timeout==3.0.0
attrs==18.1.0
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
-e git+https://github.com/Rapptz/discord.py@184c4302c7fe17a31893f0accf197c975200d233#egg=discord.py
Flask==1.0.2
future==0.16.0
idna==2.6
idna-ssl==1.0.1
itsdangerous==0.24
Jinja2==2.10
lxml==4.2.1
MarkupSafe==1.0
multidict==4.3.1
oauthlib==2.1.0
Pillow==5.4.1
pycparser==2.18
PyNaCl==1.1.2
python-twitter==3.4.2
PyYAML==3.13
requests==2.18.4
requests-oauthlib==1.0.0
six==1.11.0
urllib3==1.22
websockets==5.0.1
Werkzeug==0.14.1
yarl==1.1.1
